monitoring_creatures
The Animal Face Recognition and Image Storage System allows capturing the images of animals using a drone or a webcam, recognizing their faces, avoiding duplicates, and saving the data in a Vultr S3 bucket for safekeeping. This system offers an effective means of managing animal face imagery where uploaded duplicates are avoided in the process, thus promoting efficient usage of space and ease of access.